requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
jupyter==1.0.0
pyyaml==6.0.1
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2
imbalanced-learn==0.11.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
